Removed keyboard debug break-off
While debugging, we added the option to break of before the duration was met by pressing a key. This is now impossible (so the actual duration will always be the passed duration). Escape presses are still caught, so breaking of the experiment while the plug-in runs is possible. Also, I've streamlined the code a bit, making it less complicated that it was before.